refactor: split vsock host exec operation module
