+++ crates/vsock-host/src/exec_operation/types.rs
@@ -1,0 +1,309 @@
+use std::io;
+use std::time::Duration;
+
+use vsock_proto::{
+    ExecControlStatus, ExecOutputPolicy, ExecOutputStream, ExecTermination, ExecTimeoutPolicy,
+};
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum ExecOwnedCapturedOutput {
+    /// The stream was discarded by policy and therefore has no captured bytes.
+    Discarded,
+    /// The stream was captured, possibly with protocol-level truncation.
+    Captured {
+        /// Owned stdout or stderr bytes returned in the terminal exec result.
+        bytes: Vec<u8>,
+        /// Whether the guest truncated captured bytes to satisfy the requested
+        /// capture output policy.
+        truncated: bool,
+    },
+}
+
+/// Terminal result for a host exec operation.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct ExecOperationResult {
+    /// Terminal process state reported by the guest.
+    pub termination: ExecTermination,
+    /// Guest-reported wall-clock duration in milliseconds.
+    pub duration_ms: u32,
+    /// Stdout output state according to the requested stdout output policy.
+    pub stdout: ExecOwnedCapturedOutput,
+    /// Stderr output state according to the requested stderr output policy.
+    pub stderr: ExecOwnedCapturedOutput,
+    /// Guest-provided terminal diagnostic text.
+    ///
+    /// This may be empty when the guest has no additional diagnostic to
+    /// report for the terminal state.
+    pub diagnostic: String,
+    /// Whether the host-side bounded stream queue overflowed.
+    ///
+    /// This is separate from guest-side capture or stream truncation. When
+    /// true, some streamed output events may have been dropped even though the
+    /// terminal result was still received.
+    pub stream_overflowed: bool,
+}
+
+/// Streamed output event for a host exec operation.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct ExecOutputEvent {
+    /// Output stream that produced this event.
+    pub stream: ExecOutputStream,
+    /// Per-operation output sequence number validated by the host.
+    ///
+    /// The sequence orders stdout and stderr events together for the exec
+    /// operation.
+    pub output_seq: u32,
+    /// Output bytes emitted for this event.
+    ///
+    /// This may be empty only when the event marks stream truncation.
+    pub chunk: Vec<u8>,
+    /// Whether the guest truncated this stream to satisfy the requested stream
+    /// output policy.
+    ///
+    /// Host-side stream queue overflow is reported separately on
+    /// `ExecOperationResult::stream_overflowed`.
+    pub truncated: bool,
+}
+
+/// Request parameters for starting an exec operation.
+pub struct ExecOperationRequest<'a> {
+    /// Positive guest-side process timeout in milliseconds.
+    ///
+    /// A zero timeout is rejected before the request is sent. Use supervised
+    /// exec for operations that should not have a one-shot process timeout.
+    pub timeout_ms: u32,
+    /// Shell command to run in the guest.
+    pub command: &'a str,
+    /// Environment variables injected into the guest shell command.
+    pub env: &'a [(&'a str, &'a str)],
+    /// Whether to run the command with guest-side sudo handling.
+    pub sudo: bool,
+    /// Human-readable label used for diagnostics and logs.
+    pub label: &'a str,
+    /// Stdout output policy requested from the guest.
+    pub stdout: ExecOutputPolicy,
+    /// Stderr output policy requested from the guest.
+    pub stderr: ExecOutputPolicy,
+    /// Exit codes sent to the guest for expected-exit handling and diagnostics.
+    ///
+    /// The host still preserves the terminal state reported by the guest in
+    /// `ExecOperationResult::termination`.
+    pub expected_exit_codes: &'a [i32],
+    /// Optional bounded stdin payload written to the child and then closed.
+    pub stdin_bytes: Option<&'a [u8]>,
+    /// Optional bounded host-side output event queue override.
+    ///
+    /// `None` uses the default queue capacity when either output policy
+    /// streams, and creates no queue when neither output policy streams.
+    /// `Some` is valid only when stdout or stderr streams; zero and oversized
+    /// capacities are rejected.
+    pub stream_queue_capacity: Option<usize>,
+}
+
+/// Request parameters for a capture-only exec operation helper.
+pub struct ExecCaptureRequest<'a> {
+    /// Positive guest-side process timeout in milliseconds.
+    ///
+    /// A zero timeout is rejected before the request is sent.
+    pub timeout_ms: u32,
+    /// Shell command to run in the guest.
+    pub command: &'a str,
+    /// Environment variables injected into the guest shell command.
+    pub env: &'a [(&'a str, &'a str)],
+    /// Whether to run the command with guest-side sudo handling.
+    pub sudo: bool,
+    /// Human-readable label used for diagnostics and logs.
+    pub label: &'a str,
+    /// Maximum stdout bytes to retain in the terminal exec result.
+    pub stdout_limit_bytes: u32,
+    /// Maximum stderr bytes to retain in the terminal exec result.
+    pub stderr_limit_bytes: u32,
+    /// Exit codes sent to the guest for expected-exit handling and diagnostics.
+    pub expected_exit_codes: &'a [i32],
+    /// Optional bounded stdin payload written to the child and then closed.
+    pub stdin_bytes: Option<&'a [u8]>,
+    /// Maximum host-side wait for the terminal result after the request starts.
+    ///
+    /// This is separate from `timeout_ms`, which is the guest-side process
+    /// timeout encoded in the exec request.
+    pub wait_timeout: Duration,
+}
+
+/// Request parameters for a streaming exec operation helper.
+pub struct ExecStreamRequest<'a> {
+    /// Positive guest-side process timeout in milliseconds.
+    ///
+    /// A zero timeout is rejected before the request is sent.
+    pub timeout_ms: u32,
+    /// Shell command to run in the guest.
+    pub command: &'a str,
+    /// Environment variables injected into the guest shell command.
+    pub env: &'a [(&'a str, &'a str)],
+    /// Whether to run the command with guest-side sudo handling.
+    pub sudo: bool,
+    /// Human-readable label used for diagnostics and logs.
+    pub label: &'a str,
+    /// Stdout output policy requested from the guest.
+    ///
+    /// At least one of stdout or stderr must use a streaming output policy.
+    pub stdout: ExecOutputPolicy,
+    /// Stderr output policy requested from the guest.
+    ///
+    /// At least one of stdout or stderr must use a streaming output policy.
+    pub stderr: ExecOutputPolicy,
+    /// Exit codes sent to the guest for expected-exit handling and diagnostics.
+    pub expected_exit_codes: &'a [i32],
+    /// Optional bounded stdin payload written to the child and then closed.
+    pub stdin_bytes: Option<&'a [u8]>,
+    /// Optional host-side output event queue capacity override.
+    ///
+    /// `None` uses the default queue capacity. Zero and oversized capacities
+    /// are rejected.
+    pub stream_queue_capacity: Option<usize>,
+}
+
+/// Exec control policy for supervised host operations.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum SupervisedExecControl {
+    /// Do not register an exec-control route for the operation.
+    Disabled,
+    /// Register an exec-control route.
+    ///
+    /// When `sink` is true, the guest also exposes the bootstrap endpoint to
+    /// the child process through the process-control environment variable.
+    Enabled { sink: bool },
+}
+
+/// Request parameters for starting a supervised exec operation.
+pub struct SupervisedExecRequest<'a> {
+    /// Guest-side process timeout policy.
+    ///
+    /// `ExecTimeoutPolicy::None` lets the process run until it exits, is
+    /// cancelled, or the connection closes.
+    pub timeout: ExecTimeoutPolicy,
+    /// Shell command to run in the guest.
+    pub command: &'a str,
+    /// Environment variables injected into the guest shell command.
+    pub env: &'a [(&'a str, &'a str)],
+    /// Whether to run the command with guest-side sudo handling.
+    pub sudo: bool,
+    /// Human-readable label used for diagnostics and logs.
+    pub label: &'a str,
+    /// Stdout output policy requested from the guest.
+    pub stdout: ExecOutputPolicy,
+    /// Stderr output policy requested from the guest.
+    pub stderr: ExecOutputPolicy,
+    /// Exit codes that should be marked expected in the guest-side exec request.
+    pub expected_exit_codes: &'a [i32],
+    /// Optional bounded stdin payload written to the child and then closed.
+    pub stdin_bytes: Option<&'a [u8]>,
+    /// Optional exec-control route for this supervised operation.
+    pub control: SupervisedExecControl,
+    /// Optional bounded host-side output event queue override.
+    ///
+    /// `None` uses the default queue capacity when either output policy
+    /// streams, and creates no queue when neither output policy streams.
+    /// `Some` is valid only when stdout or stderr streams; zero and oversized
+    /// capacities are rejected.
+    pub stream_queue_capacity: Option<usize>,
+    /// Maximum time to wait for the guest `exec_started` acknowledgement.
+    ///
+    /// If this elapses after the start frame is written, the host sends
+    /// `MSG_EXEC_CANCEL` for the operation before returning a timeout error.
+    /// If that cancel frame cannot be written within the bounded fallback
+    /// window, the connection is poisoned because the guest process state is
+    /// no longer known.
+    ///
+    /// A successful start-timeout cancellation still abandons terminal proof
+    /// for this operation, so the connection should not be reused for later
+    /// normal operations.
+    pub start_timeout: Duration,
+}
+
+/// Host-side acknowledgement for a delivered exec-control request.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct ExecControlAck {
+    /// Sequence number of the supervised exec operation that received control.
+    pub target_seq: u32,
+    /// Control message identifier echoed by the guest.
+    pub message_id: String,
+}
+
+/// Guest-side terminal status for an exec-control request.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct ExecControlGuestStatus {
+    /// Guest-reported control status other than delivered.
+    pub status: ExecControlStatus,
+    /// Guest-provided diagnostic text for the status.
+    ///
+    /// This may be empty when the guest has no additional diagnostic to
+    /// report.
+    pub diagnostic: String,
+}
+
+/// Terminal guest response for an exec-control request.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum ExecControlOutcome {
+    /// The guest delivered the request to the control sink and accepted it.
+    Delivered(ExecControlAck),
+    /// The guest responded with a control status other than delivered.
+    GuestStatus(ExecControlGuestStatus),
+    /// The guest returned an error response for the pending control request.
+    GuestError(String),
+}
+
+impl ExecControlOutcome {
+    /// Return the delivered acknowledgement or convert guest failures to an error.
+    ///
+    /// Only `ExecControlOutcome::Delivered` yields `Ok`. Guest statuses and
+    /// guest error responses become `io::Error` values.
+    pub fn into_ack(self) -> io::Result<ExecControlAck> {
+        match self {
+            Self::Delivered(ack) => Ok(ack),
+            Self::GuestStatus(status) => {
+                Err(exec_control_status_error(status.status, &status.diagnostic))
+            }
+            Self::GuestError(message) => Err(io::Error::other(message)),
+        }
+    }
+}
+
+pub(in crate::exec_operation) fn exec_control_status_error(
+    status: ExecControlStatus,
+    diagnostic: &str,
+) -> io::Error {
+    let message = if diagnostic.is_empty() {
+        default_exec_control_status_message(status).to_owned()
+    } else {
+        diagnostic.to_owned()
+    };
+    io::Error::new(exec_control_status_error_kind(status), message)
+}
+
+fn exec_control_status_error_kind(status: ExecControlStatus) -> io::ErrorKind {
+    match status {
+        ExecControlStatus::Delivered => io::ErrorKind::Other,
+        ExecControlStatus::Inactive => io::ErrorKind::NotFound,
+        ExecControlStatus::NonceMismatch => io::ErrorKind::PermissionDenied,
+        ExecControlStatus::Unsupported => io::ErrorKind::Unsupported,
+        ExecControlStatus::Rejected => io::ErrorKind::PermissionDenied,
+        ExecControlStatus::SinkUnavailable => io::ErrorKind::NotConnected,
+        ExecControlStatus::SinkTimeout => io::ErrorKind::TimedOut,
+        ExecControlStatus::QueueFull => io::ErrorKind::WouldBlock,
+        ExecControlStatus::SinkError => io::ErrorKind::BrokenPipe,
+    }
+}
+
+fn default_exec_control_status_message(status: ExecControlStatus) -> &'static str {
+    match status {
+        ExecControlStatus::Delivered => "exec control request delivered",
+        ExecControlStatus::Inactive => "exec operation is not active",
+        ExecControlStatus::NonceMismatch => "exec operation nonce mismatch",
+        ExecControlStatus::Unsupported => "exec control is not supported by this operation",
+        ExecControlStatus::Rejected => "exec control request rejected",
+        ExecControlStatus::SinkUnavailable => "exec control sink is not connected",
+        ExecControlStatus::SinkTimeout => "exec control sink timed out",
+        ExecControlStatus::QueueFull => "exec control queue is full",
+        ExecControlStatus::SinkError => "exec control sink error",
+    }
+}
